Fixed issues where Marshal.AllocCoTaskMem was called with uint and int* arguments. Made require tags bold.

+++ src/OpenTK.Graphics/OpenGL/Compatibility/GL.Overloads.cs
@@ -40312,8 +40312,8 @@ namespace OpenTK.Graphics.OpenGL.Compatibility
                 fixed (uint* counterDataTypeEnum_ptr = &counterDataTypeEnum)
                 fixed (ulong* rawCounterMaxValue_ptr = &rawCounterMaxValue)
                 {
-                    var counterName_ptr = (byte*)Marshal.AllocCoTaskMem(counterNameLength);
-                    var counterDesc_ptr = (byte*)Marshal.AllocCoTaskMem(counterDescLength);
+                    var counterName_ptr = (byte*)Marshal.AllocCoTaskMem((int)counterNameLength);
+                    var counterDesc_ptr = (byte*)Marshal.AllocCoTaskMem((int)counterDescLength);
                     GetPerfCounterInfoINTEL(queryId, counterId, counterNameLength, counterName_ptr, counterDescLength, counterDesc_ptr, counterOffset_ptr, counterDataSize_ptr, counterTypeEnum_ptr, counterDataTypeEnum_ptr, rawCounterMaxValue_ptr);
                     counterDesc = Marshal.PtrToStringUTF8((IntPtr)counterDesc_ptr)!;
                     Marshal.FreeCoTaskMem((IntPtr)counterDesc_ptr);
@@ -40353,7 +40353,7 @@ namespace OpenTK.Graphics.OpenGL.Compatibility
                 fixed (uint* noInstances_ptr = &noInstances)
                 fixed (uint* capsMask_ptr = &capsMask)
                 {
-                    var queryName_ptr = (byte*)Marshal.AllocCoTaskMem(queryNameLength);
+                    var queryName_ptr = (byte*)Marshal.AllocCoTaskMem((int)queryNameLength);
                     GetPerfQueryInfoINTEL(queryId, queryNameLength, queryName_ptr, dataSize_ptr, noCounters_ptr, noInstances_ptr, capsMask_ptr);
                     queryName = Marshal.PtrToStringUTF8((IntPtr)queryName_ptr)!;
                     Marshal.FreeCoTaskMem((IntPtr)queryName_ptr);
@@ -40367,7 +40367,7 @@ namespace OpenTK.Graphics.OpenGL.Compatibility
                 fixed (uint* noInstances_ptr = &noInstances)
                 fixed (uint* capsMask_ptr = &capsMask)
                 {
-                    var queryName_ptr = (byte*)Marshal.AllocCoTaskMem(queryNameLength);
+                    var queryName_ptr = (byte*)Marshal.AllocCoTaskMem((int)queryNameLength);
                     GetPerfQueryInfoINTEL(queryId, queryNameLength, queryName_ptr, dataSize_ptr, noCounters_ptr, noInstances_ptr, capsMask_ptr);
                     queryName = Marshal.PtrToStringUTF8((IntPtr)queryName_ptr)!;
                     Marshal.FreeCoTaskMem((IntPtr)queryName_ptr);
